move ExecutionPaths out of mill.exec
