+++ kotlinlib/test/src/mill/kotlinlib/js/KotlinJsKotestModuleTests.scala
@@ -3,7 +3,7 @@ package kotlinlib.js
 
 import mill.api.ExecResult
 import mill.define.Discover
-import mill.exec.ExecutionPaths
+import mill.define.ExecutionPaths
 import mill.testkit.{TestBaseModule, UnitTester}
 import utest.{TestSuite, Tests, assert, test}
 
@@ -45,7 +45,7 @@ object KotlinJsKotestModuleTests extends TestSuite {
         eval.apply(command): @unchecked
 
       val xmlReport =
-        ExecutionPaths.resolveDestPaths(eval.outPath, command).dest / "test-report.xml"
+        ExecutionPaths.resolve(eval.outPath, command).dest / "test-report.xml"
 
       assert(
         os.exists(xmlReport),
