+++ scalalib/test/src/mill/scalalib/CrossVersionTests.scala
@@ -1,7 +1,7 @@
 package mill.scalalib
 
 import mill.define.Discover
-import mill.exec.ExecutionPaths
+import mill.define.ExecutionPaths
 import mill.main.TokenReaders.*
 import mill.testkit.UnitTester
 import mill.testkit.TestBaseModule
@@ -147,7 +147,7 @@ object CrossVersionTests extends TestSuite {
         // Escape-sequence formatting isn't working under bare Windows
         val expectedDepsTree = tree
         val depsTree =
-          os.read(ExecutionPaths.resolveDestPaths(eval.execution.workspace, mod.ivyDepsTree(IvyDepsTreeArgs())).log)
+          os.read(ExecutionPaths.resolve(eval.execution.workspace, mod.ivyDepsTree(IvyDepsTreeArgs())).log)
         assert(depsTree == expectedDepsTree)
       }
     }
